Create the XML of sample

This function generates XML files for each sample.

Usage

specimen_XML(
  excel_file = NULL,
  excel_curators = NULL,
  excel_relation = NULL,
  excel_sampler = NULL,
  specimen_uuids = NULL,
  root_dir = NULL
)

Arguments

excel_file

A tibble that contains the columns present in the 'SpecimenInfo' sheet of the spreadsheet template.

excel_curators

A tibble that contains the columns present in the 'CuratorsInfo' sheet of the spreadsheet template.

excel_relation

A tibble that contains the columns present in the 'RelationInfo' sheet of the spreadsheet template.

excel_sampler

A tibble that contains the columns present in the 'SamplersInfo' sheet of the spreadsheet template.

specimen_uuids

A character that contains a number of strings UUIDs equal to the number of samples present in the SpecimenInfo' sheet of the spreadsheet template.

root_dir

A string with the path to the folder where the ttl and XML files will be written.

Author

Alessandro Oggioni, phD (2023) oggioni.a@cnr.it